// src/AdsetManagementWeb/src/app/api/models/UpdateVehicleRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OtherOptionsRequest } from './OtherOptionsRequest';
export type UpdateVehicleRequest = {
    imagens?: Array<string> | null;
    marca: string;
    modelo: string;
    ano: string;
    placa: string;
    km?: number | null;
    cor: string;
    preco: number;
    otherOptions?: OtherOptionsRequest;
    pacoteICarros?: string | null;
    pacoteWebMotors?: string | null;
};